Moved interfaces to Application
