--- Application/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Application
-{
-    public class Class1
-    {
-
-    }
-}
