+++ Application/IEventService.cs
@@ -1,4 +1,6 @@
-namespace Entities;
+using Entities;
+
+namespace Application;
 
 public interface IEventService
 {
