+++ CleanArchitectureBlazor/CleanArchitectureBlazor/CleanArchitectureBlazor.csproj
@@ -9,6 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Application\Application.csproj" />
     <ProjectReference Include="..\..\Entities\Entities.csproj" />
     <ProjectReference Include="..\..\Infrastructure\Infrastructure.csproj" />
     <ProjectReference Include="..\CleanArchitectureBlazor.Client\CleanArchitectureBlazor.Client.csproj" />
