+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Dashboard.razor
@@ -1,5 +1,6 @@
 @page "/dashboard"
 @using Entities
+@using Application
 @inject IEventRepository EventService
 @inject IShiftRepository ShiftService
 @inject IStaffRepository StaffService
