+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/EventDetails.razor
@@ -1,5 +1,6 @@
 @page "/events/details/{EventId:int}"
 @using Entities
+@using Application
 @inject IEventRepository EventService
 @inject IShiftRepository ShiftService
 @inject NavigationManager Navigation
