+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/EventEdit.razor
@@ -1,5 +1,6 @@
 @page "/events/edit/{EventId:int}"
 @using Entities
+@using Application
 @inject IEventRepository EventRepository
 @inject IShiftRepository ShiftService
 @inject IEventService EventService
