+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Events.razor
@@ -1,5 +1,6 @@
 @page "/events"
 @using Entities
+@using Application
 @inject IEventRepository EventService
 @inject IShiftRepository ShiftService
 @inject NavigationManager Navigation
