+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/ShiftDetails.razor
@@ -1,5 +1,6 @@
 @page "/shifts/details/{ShiftId:int}"
 @using Entities
+@using Application
 @inject IShiftRepository ShiftService
 @inject IStaffAssignmentRepository StaffAssignmentService
 @inject NavigationManager Navigation
