+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/ShiftEdit.razor
@@ -1,5 +1,6 @@
 @page "/shifts/edit/{ShiftId:int}"
 @using Entities
+@using Application
 @inject IShiftRepository ShiftService
 @inject IStaffRepository StaffService
 @inject IStaffAssignmentRepository StaffAssignmentService
