+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Shifts.razor
@@ -1,5 +1,6 @@
 @page "/shifts"
 @using Entities
+@using Application
 @inject IShiftRepository ShiftService
 @inject IEventRepository EventService
 @inject NavigationManager Navigation
