+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Staff.razor
@@ -1,4 +1,5 @@
 @page "/staff"
+@using Application
 @inject IStaffRepository StaffService
 @inject NavigationManager Navigation
 @rendermode InteractiveServer
