+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/StaffDetails.razor
@@ -1,4 +1,5 @@
 @page "/staff/details/{StaffId:int}"
+@using Application
 @using Entities
 @inject IStaffRepository StaffService
 @inject IStaffAssignmentRepository StaffAssignmentService
