+++ Infrastructure/Infrastructure.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Application\Application.csproj" />
     <ProjectReference Include="..\Entities\Entities.csproj" />
   </ItemGroup>
